Major structural changes & refactoring
Class is no longer static; Object interfaces are injected to avoid rewriting / recompiling this class. Code in conditional blocks within program loop are abstracted to 'execute' methods. TextExtractor logic is now abstracted to TextExtractorManager class to return a TextExtractorResult to be used in this class. Display methods are abstracted: System.out.println statements are abstracted to a display method & the display messages are now located in a utility Message class.